// src/app/core/models/farm-list.ts
export interface FarmList {
    firstName?: string;
	lastName?: string;
	ghlUserId?: string;
	suffix?: string;
	email?: string;
	phone?: string;
	roleId?: string;
	roleName?: string;
	farmId?: string;
	userName?: string;
	tenantId?: string;
	isActive?: boolean;
	isTenantAdmin?: boolean;
	jobTitle?: string;
	jobTitleName?: string;
	[key: string]: any;
}